% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteins.R
\docType{class}
\name{UniProtProtein-class}
\alias{.UniProtProtein}
\alias{UniProtProtein}
\alias{UniProtProtein-class}
\alias{UniProtProteinIsoforms}
\alias{UniProtProteins}
\alias{show,UniProtProtein-method}
\title{A Protein fetched from UniProt REST API}
\usage{
\S4method{show}{UniProtProtein}(object)

UniProtProtein(accession)

UniProtProteinIsoforms(accession)

UniProtProteins(accessions)
}
\arguments{
\item{object}{An instance of class \code{UniProtProtein}.}

\item{accession}{A \code{character} contain one of multiple
accession numbers. If more than one are provided, a
\code{list} of \code{UniProtProtein} objects is returned.}

\item{accessions}{A \code{character} with on or more accession numbers.}
}
\description{
A Protein fetched from UniProt REST API

Method to show a \code{UniProtProtein} object.

Constructor for \code{UniProtProtein} objects

Retrieves all isoforms for a given accession number

This constructor sends a unique message to the UniProt REST server
containing all accession numbers, as opposed to
\code{UniProtProtein}, that sends one message per accession
number.
}
\section{Slots}{

\describe{
\item{\code{Data}}{A \code{list} containing all the data.}
}}
\examples{
library("UniProt.REST")
ac <- "P27361"
acs <- c("P27361", "P38398")
UniProtProtein(ac)
UniProtProteins(acs) ## works also as UniProtProtein(ac)
UniProtProteinIsoforms(ac)
}

